messaging complete

+++ app/Http/Controllers/MailController.php
@@ -2,8 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\MailApiRequest;
+use App\Http\Requests\MailRequest;
 use App\Mail;
+use App\User;
 use Illuminate\Http\Request;
 
 class MailController extends Controller
@@ -11,50 +12,77 @@ class MailController extends Controller
 
     public function index(Request $request)
     {
+        $page = "inbox";
         $mails = $request->user()->incomingMail()
             ->orderBy('created_at', 'desc')
             ->with('sender')
             ->paginate(15);
         $items = json_encode($mails->items());
-        return view('mail.index', compact('mails', 'items'));
+        return view('mail.inbox', compact('mails', 'items', 'page'));
     }
 
-    public function sentIndex()
+    public function sentIndex(Request $request)
     {
-        return view('mail.index');
+        $page = "sent";
+        $mails = $request->user()->outgoingMail()
+            ->orderBy('created_at', 'desc')
+            ->with('receiver')
+            ->paginate(15);
+        return view('mail.sent', compact('mails', 'page'));
     }
 
-    public function create()
+    public function create(Request $request, $email = null)
     {
-        return view('mail.create');
+        $page = "create";
+        if($email != null)
+        {
+            $request->session()->flash('email', $email);
+        }
+        return view('mail.create', compact('page'));
     }
 
-    public function store()
+    public function forward(Request $request)
     {
+        $page = "forward";
+        $mail = Mail::find($request->mailId);
+        $forwardMessage = 'sent by' . $mail->sender()->first()->email . "\n\n\"" . $mail->message . "\"";
+        $request->session()->flash('message', $forwardMessage);
+        return view('mail.create', compact('page'));
+    }
 
+    public function store(MailRequest $mailRequest)
+    {
+        $receiver = User::where('email', $mailRequest->to)->first();
+        $mail = new Mail([
+            "subject" => $mailRequest->subject,
+            "message" => $mailRequest->message,
+            "read" => 0,
+            "favorite" => 0,
+        ]);
+        $mail->sender()->associate($mailRequest->user());
+        $mail->receiver()->associate($receiver);
+        $mail->save();
+        return redirect('/mail');
     }
 
     public function show(Mail $mail, Request $request)
     {
+        $page = "show";
         if($mail->receiver_id != $request->user()->id){
             return back();
         }
         $mail->setRead(true);
-        return view('mail.show', compact('mail'));
+        return view('mail.show', compact('mail', 'page'));
     }
 
-    public function destroy(Mail $mail)
+    public function destroy(Mail $mail, Request $request)
     {
-        $mail->delete();
-    }
-
-    public function destroyAll(Request $request)
-    {
-        $mails = Mail::find($request->checked);
-        foreach($mails as $mail)
+        if($mail->receiver_id != $request->user()->id)
         {
-            $mail->delete();
+            return back();
         }
+        $mail->delete();
+        return redirect('/mail');
     }
 
     public function getUserNotifications(Request $request)
@@ -63,10 +91,11 @@ class MailController extends Controller
             ->where('read', 0)
             ->orderBy('created_at', 'desc')
             ->with('sender')
+            ->take(5)
             ->get();
     }
 
-    public function mailApi(MailApiRequest $request)
+    public function mailApi(Request $request)
     {
         $mails = Mail::find($request->checked);
         foreach($mails as $mail)
